Refactor hit and sunk functionality on ships
Previously the hit functionality on ships would have required looping
through the board cells to check for a specific ship, then checking if
that cell index matched that of the cell that was attacked. Now the
markHit function on ships simply takes the coordinate that was attacked
and adds that to the (initially empty) hit array.